requirement.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
Naked==0.1.31
PyQt5==5.13.0
PyQt5-sip==12.8.0
pyqt5-tools==5.13.0.1.5
python-dotenv==0.14.0
pytube3==9.6.4
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
tqdm==4.48.0
typing-extensions==3.7.4.2
urllib3==1.25.10
